style: Change webhooks context modules to match the naming convention
